# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de prueba NITSNETS",
        "description": "Esta es una primera versión de la API",
        "contact": "jorgepastor9120@gmail.com",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Mostrar el listado de usuarios",
                "description": "Mostramos el listado de los regitros solicitados.",
                "operationId": "b343fe695d1beaa03b7deb8449dd1b7c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se ha podido cargar el listado."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Registra un usuario",
                "description": "Registrar un usuario.",
                "operationId": "71d86dd873f9e567415d5da5929a0dea",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "El usuario ha sido creado"
                    },
                    "404": {
                        "description": "No se ha podido registrar el usuario."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualiza un usuario",
                "description": "Actualizar un registro.",
                "operationId": "e27513375b71e3b3dcc043da6f62129c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Actualiza un usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "example": "Prueba"
                                    },
                                    "email": {
                                        "type": "text",
                                        "example": "prueba@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": ""
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido actualizar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Elimina un usuario",
                "description": "Elimina un registro.",
                "operationId": "ae967f8cf847f7d2ad9eba19833f58cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Elimina un usuario"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido eliminar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "name": "authorization",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/api/v1/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}